cicd-cloud-function/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asttokens==2.2.1
backcall==0.2.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
fastapi==0.100.0
google-api-core==2.11.1
google-auth==2.22.0
google-cloud==0.34.0
google-cloud-core==2.3.3
google-cloud-pubsub==2.18.0
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpc-google-iam-v1==0.12.6
grpcio==1.56.0
grpcio-status==1.56.0
gunicorn==20.1.0
h11==0.14.0
httptools==0.6.0
idna==3.4
ipykernel==6.24.0
ipython==8.14.0
jedi==0.18.2
joblib==1.3.1
jupyter_client==8.3.0
jupyter_core==5.3.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.25.1
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.9.1
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.23.4
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.0.3
pydantic_core==2.3.0
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
pyzmq==25.1.0
requests==2.31.0
rsa==4.9
scikit-learn==1.2.1
scipy==1.11.1
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.27.0
threadpoolctl==3.2.0
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
